+++ activerecord/lib/active_record/connection_adapters/abstract/schema_creation.rb
@@ -30,7 +30,7 @@ module ActiveRecord
           def visit_ColumnDefinition(o)
             sql_type = type_to_sql(o.type, o.limit, o.precision, o.scale)
             column_sql = "#{quote_column_name(o.name)} #{sql_type}"
-            add_column_options!(column_sql, column_options(o)) unless o.primary_key?
+            add_column_options!(column_sql, column_options(o)) unless o.type == :primary_key
             column_sql
           end
 
@@ -65,6 +65,8 @@ module ActiveRecord
             column_options[:column] = o
             column_options[:first] = o.first
             column_options[:after] = o.after
+            column_options[:auto_increment] = o.auto_increment
+            column_options[:primary_key] = o.primary_key
             column_options
           end
 
@@ -88,6 +90,9 @@ module ActiveRecord
             end
             if options[:auto_increment] == true
               sql << " AUTO_INCREMENT"
+            end
+            if options[:primary_key] == true
+              sql << " PRIMARY KEY"
             end
             sql
           end